Add PayPalWebCheckoutVaultExperienceContext type.
